fix https://github.com/umjammer/vavi-image-sandbox/security/code-scanning/14

+++ src/main/java/vavix/awt/image/resample/enlarge/smilla/FractTab.java
@@ -148,7 +148,7 @@ class FractTab {
             int s1, s2, hh;
             rr += (int) (randG.randL() >> 5);
             s1 = rr & RandTabMask;
-            rr += (randG.randL() >> 5);
+            rr += (int) (randG.randL() >> 5);
             s2 = rr & RandTabMask;
 
             hh = randTab[s1];
